+++ src/v/datalake/BUILD
@@ -126,8 +126,8 @@ redpanda_cc_library(
     deps = [
         ":logger",
         "//src/v/base",
-        "//src/v/cloud_io:auth_refresh_bg_op",
         "//src/v/cloud_roles",
+        "//src/v/cloud_roles:auth_refresh_bg_op",
         "//src/v/cloud_roles:types",
         "//src/v/cloud_storage_clients",
         "//src/v/config",
