+++ src/v/datalake/coordinator/BUILD
@@ -75,7 +75,6 @@ redpanda_cc_library(
         ":iceberg_file_committer",
         ":iceberg_snapshot_remover",
         "//src/v/base",
-        "//src/v/cloud_io:auth_refresh_bg_op",
         "//src/v/cloud_storage_clients",
         "//src/v/config",
         "//src/v/datalake:catalog_schema_manager",
@@ -96,7 +95,6 @@ redpanda_cc_library(
         ":file_committer",
         ":state_update",
         ":stm",
-        "//src/v/cloud_io:auth_refresh_bg_op",
         "//src/v/cloud_io:remote",
         "//src/v/cloud_roles",
         "//src/v/cluster",
